feat: extract suggestion utilities
