+++ public/index.html
@@ -13,6 +13,7 @@
     <h2>Suggestions</h2>
     <ul id="suggestion-list"></ul>
   </section>
+  <script src="suggestions.js"></script>
   <script src="app.js"></script>
 </body>
 </html>
